Add struct node for linked list and insertLast()
Struct node stores the value of the element and a pointer to the next
and previous element in the doubly linked list. The insertLast()
function takes in the head of the list and inserts an element at the end
of the list.

Adds utility to create a doubly linked list and insert an element at the
end of the list.