Add printList to print list from head or tail
The printList(struct node *head) function takes in struct node pointer
and prints the list in either normal order or reverse order dependent
on if the pointer that is passed in is the head or the tail of the list.
The function checks if the previous or next element is null to determine
if the head or the tail of the list was passed in before printing.

Access to a printList function allows the list to be printed for testing
the functionality of the utility functions.