Add struct node **tail to insertLast function
In the insertLast function the double pointer struct node **tail
was added as a parameter. The value of this pointer is set to the last
element in the list. The function is called with the address of a
single pointer struct node *tail with the double pointer allowing the
value of the pointer *tail to be changed.

The addition of the extra parameter allows for the tail of the list to
be updated at the same time that the head of the list is updated.