Change insertLast by using tail for insertion
Change the insertLast function by using the value of the tail to
insert a value at the end of the list to allow for the operation
to be done in O(1) time instead of O(n) time.

Using tail to insert the final value saves time as it is not necessary
to iterate to the end of the list.