Add tail and head operability for get and sizeList
Add the ability for the head or the tail of the list to be passed into
get and sizeList. Using if statements it is determined if the head
or the tail of the list was the variable passed into the function.
The condition if(temp-> prev == NULL) evaluates to true if the
variable passed in was head and otherwise the variable passed in
was the tail of the list.

The ability to use the head or the tail of the list allows for more
options when manipulating the doubly linked list.