Add removeElement function and testing in main
The removeElement function is a function that takes in the head of
a list and the index of an element to be removed. The function removes
the element by pointing the previous element in the list to the next
element and the next element to the previous element if the element to
be removed is not the last element. If the element to be removed is the
last element the next element of the previous element is null.
Additional testing of the functions in main using printf to segment
the testing of the functions.

The removeElement function allows for removal of certain elements
from a list. This functionality is an important feature in a doubly
linked list.