Add reverseList to reverse elements in list
The reverseList function takes in the head of a list as a parameter
and recursively iterates through the list swapping elements until
the elements of the list are reversed.

The reverseList function allows for the elements in the list to be
placed in reverse order and is useful for demonstrating recursion.